refactor(statistics): break down monolithic page into reusable components
Decompose the large statistics page (1000+ lines) into smaller, maintainable components and utilities:

- Extract chart components: DrillPerformanceChart, WeeklyTestChart
- Create UI components: StatisticCard, StatisticsOverview, RecentActivities, WeeklyTestsTable
- Add utility modules: chartHelpers, dateFormatters, statisticsCalculators, weeklyDataProcessors
- Introduce dedicated statisticsStore for centralized state management
- Define proper TypeScript interfaces for Chart and Statistics domains
- Consolidate constants in statistics.ts

This improves code maintainability, testability, and reusability while preserving all existing functionality.
