+++ src/lib/utils/dateFormatters.ts
@@ -1,0 +1,45 @@
+/**
+ * Formats a date to a short string (e.g., "Dec 25")
+ */
+export function formatDate(date: Date): string {
+  return date.toLocaleDateString('en-US', {
+    month: 'short',
+    day: 'numeric'
+  });
+}
+
+/**
+ * Formats seconds to MM:SS string
+ */
+export function formatTime(seconds: number): string {
+  if (!isFinite(seconds) || seconds < 0) return '00:00';
+  const minutes = Math.floor(seconds / 60);
+  const remainingSeconds = Math.floor(seconds % 60);
+  return `${String(minutes).padStart(2, '0')}:${String(remainingSeconds).padStart(2, '0')}`;
+}
+
+/**
+ * Formats date for weekly display (DD/MM/YYYY)
+ */
+export function formatWeeklyDate(dateStr: string): string {
+  try {
+    return new Date(dateStr).toLocaleDateString('en-GB', {
+      day: '2-digit',
+      month: '2-digit',
+      year: 'numeric'
+    });
+  } catch {
+    return 'Invalid Date';
+  }
+}
+
+/**
+ * Normalizes date string to YYYY-MM-DD format
+ */
+export function normalizeDate(dateStr: string): string {
+  try {
+    return new Date(dateStr).toISOString().split('T')[0];
+  } catch {
+    return dateStr;
+  }
+}
